support: add qurlew-shell login script
This script can be used as a user login shell.

When user logs in, it will allow to change the URL of the qurlew
instance.

First, the script must be installed.

	$ sudo install -m 755 support/qurlew-shell /usr/bin/

Then a qurlew user must be added. The command varies depending on the
distribution.

The following command creates a qurlew user with the current user
UID/GID.

	$ sudo bash -c "useradd --non-unique --gid ${GROUPS[0]} --uid $UID --no-create-home --shell /usr/bin/qurlew-shell qurlew"

Update qurlew password to be able to log in as qurlew user.

	$ sudo passwd qurlew
	New password:
	Retype new password:
	passwd: password updated successfully

Here is the result in /etc/passewd.

	$ grep qurlew /etc/passwd
	qurlew:x:1000:1000::/home/qurlew:/usr/bin/qurlew-shell

Now, log in as qurlew user.

	$ su qurlew
	Password:
	http://www.savoirfairelinux.com
	^CInterrupt

Or, 'run' a command as qurlew.

	$ su qurlew -c http://www.buildroot.org
	Password: